prepare example functions
