+++ jest.config.js
@@ -1,0 +1,13 @@
+module.exports = {
+  verbose: true,
+  collectCoverage: true,
+  testEnvironment: "node",
+  coverageThreshold: {
+    global: {
+      branches: 80,
+      functions: 80,
+      lines: 80,
+      statements: 80,
+    },
+  },
+}
