+++ src/_shared/array.all.ts
@@ -1,0 +1,3 @@
+export const all = <T>(iterable: Iterable<T>, predict: (e: T, index: number) => boolean): boolean => {
+  return Array.from(iterable).reduce<boolean>((prev, e, i) => prev && predict(e, i), true)
+}
