+++ src/_shared/array.count.ts
@@ -1,0 +1,3 @@
+export const count = <T>(array: T[], predict: (element: T) => boolean): number => {
+  return array.reduce((prev, element) => (predict(element) ? prev + 1 : prev), 0)
+}
