+++ src/_shared/array.excludeNull.ts
@@ -1,0 +1,3 @@
+export const excludeNull = <T>(array: (T | null | undefined)[]): T[] => {
+  return array.filter((e: T | null | undefined): e is T => !!e)
+}
