+++ src/_shared/array.mapRepeated.ts
@@ -1,0 +1,3 @@
+export const mapRepeated = <TResult>(count: number, map: (index: number) => TResult): TResult[] => {
+  return Array.from({ length: count }).map((_, index) => map(index))
+}
