+++ src/_shared/array.max.ts
@@ -1,0 +1,9 @@
+import { sortByNumberValue } from './array.sortByNumberValue';
+
+export const max = <T>(array: T[], getValue: (element: T) => number): T | undefined => {
+  if (array.length === 0) {
+    return undefined
+  }
+  const sorted = array.sort((a, b) => sortByNumberValue(getValue(a), getValue(b), 'desc'));
+  return sorted[0];
+}
