+++ src/_shared/array.sortByNumberValue.ts
@@ -1,0 +1,3 @@
+import sortByComparableValue from './array.sortByComparableValue';
+
+export const sortByNumberValue = (a: number, b: number, order: 'asc' | 'desc') => sortByComparableValue(a, b, order);
