+++ src/_shared/array.sortByStringValue.ts
@@ -1,0 +1,3 @@
+import sortByComparableValue from './array.sortByComparableValue';
+
+export const sortByStringValue = (a: string, b: string, order: 'asc' | 'desc') => sortByComparableValue(a, b, order);
