+++ src/_shared/array.unique.ts
@@ -1,0 +1,3 @@
+export const unique = <T>(array: T[]): T[] => {
+  return Array.from(new Set(array))
+}
